Added twitter functionality for interactive map and control card reader
The map now contains a twitter button at the mosaic area.
Twitter functionality has been added (needs testing on robot)
A nod using the control card has been added (needs testing on robot and an idea for the collect command)
The pausing only pauses the navigation and not the scheduler.